fix solials list
