+++ src/components/BackButton/BackButton.jsx
@@ -1,4 +1,4 @@
-import { useHistory } from 'react-router';
+import { useHistory } from 'react-router-dom';
 import { Button } from 'react-bootstrap';
 
 const BackButton = () => {
