+++ src/components/SocialsList/SocialsList.jsx
@@ -1,13 +1,32 @@
+import { useEffect } from 'react';
+import { useDispatch, useSelector } from 'react-redux';
+import { useAppContext } from '../../contexts/App.context';
 import { Spinner } from 'react-bootstrap';
 
 import SocialPreview from '../SocialPreview/SocialPreview';
 
-const SocialsList = ({ list, loading, error }) => {
+import { loadAllSocials, clearAllSocials } from '../../ducks/socials/all';
+import selector from './SocialsList.selector';
+
+const SocialsList = () => {
+  const dispatch = useDispatch();
+
+  const { appId } = useAppContext();
+  const { all, loading } = useSelector(selector);
+
+  useEffect(() => {
+    dispatch(loadAllSocials(appId));
+
+    return () => {
+      dispatch(clearAllSocials());
+    };
+  }, [dispatch, appId]);
+
   return loading ? (
     <Spinner animation='border' />
   ) : (
     <>
-      {list.map(({ id, name, imgUrl, isSetting }) => (
+      {all.map(({ id, name, imgUrl, isSetting }) => (
         <SocialPreview
           key={id}
           id={id}
